Mods to avoid Visual Studio warnings in sim360.

+++ XCOM-I/sim360-source/simulate/smop.c
@@ -3,6 +3,9 @@
 **
 **	Author: Daniel Weaver
 */
+/*
+ * 2024-08-18 RSB	Some casts in a printf to avoid Visual Studio warnings.
+ */
 
 #include <math.h>
 #include <float.h>
@@ -79,8 +82,8 @@ cnv_ibm32(double n)
 void
 print_float(XPL_FLOAT *v)
 {
-	printf("%c %04x %016lx %-8lg", v->sign ? '-' : '+', v->exp & 0xffff,
-		v->fraction, cnv_double(v));
+	printf("%c %04x %016lx %-8lg", v->sign ? '-' : '+', (long) v->exp & 0xffff,
+		(long) v->fraction, cnv_double(v));
 }
 
 /*
